+++ resources/js/vendor.js
@@ -1,0 +1,7 @@
+try {
+    window.$ = window.jQuery = require('jquery');
+    require('bootstrap');
+    require('feather-icons').replace();
+    window._ = require('lodash');
+    window.axios = require('axios');
+} catch (e) { }