+++ src/Http/Controllers/Dashboard.php
@@ -4,7 +4,7 @@ namespace Origami\Admin\Http\Controllers;
 
 class Dashboard extends AdminController {
 
-    public function index()
+    public function __invoke()
     {
         return view('admin.dashboard.index');
     }