+++ webpack.mix.js
@@ -1,26 +1,10 @@
-let mix = require('laravel-mix');
+const mix = require('laravel-mix');
 
-mix.options({ publicPath: 'public/assets' });
-mix.setResourceRoot('/assets/');
+mix.options({ publicPath: 'public' });
 
-/*
- |--------------------------------------------------------------------------
- | Mix Asset Management
- |--------------------------------------------------------------------------
- |
- | Mix provides a clean, fluent API for defining some Webpack build steps
- | for your Laravel application. By default, we are compiling the Sass
- | file for the application as well as bundling up all the JS files.
- |
- */
+mix.version();
 
-mix.sass('resources/assets/sass/admin.scss', 'public/assets/css')
-    .js('resources/assets/js/admin.js', 'public/assets/js');
+mix.js('resources/js/vendor.js', 'public/js');
 
-// mix.browserSync({
-//     proxy: 'admin.test',
-//     files: [
-//         'public/**/*.js',
-//         'public/**/*.css'
-//     ]
-// });
+mix.sass('resources/sass/admin.scss', 'public/css')
+    .js('resources/js/admin.js', 'public/js');