+++ changelog.md
@@ -1,3 +1,12 @@
+## 3.0.0
+- Renamed admin.url config to admin.path
+- Removed aurl helper (use admin_url).
+- Renamed admin-assets publishing tag to admin-resources.
+- Renamed admin-public publishing tag to admin-assets.
+- New public/vendor/admin publishing path for assets.
+- New admin:: namespace and views/vendor/admin publishing path for package views.
+- Removed summernote (TODO).
+
 ## 2.0.0
 - New major version with custom Bootstrap 4.0 template.
 - Removed bootbox javascript library.